style: golangci-lint clean — 0 issues, from 99
The reported baseline was never 91. golangci-lint defaults to
max-issues-per-linter=50, so revive was CAPPED and the real total was 99; the
first exclusion I added removed 50 real issues and simply let a different 50
surface at the same number. Everything below is measured uncapped.

CONFIG (42 issues): the "should have comment or be unexported" exclusion list was
incomplete, not the code. It already covered internal/core, store/badger and
store/tiered; index, storeutil, memory and sharded are the SAME two categories
and were never added. Now stated as two rules with their reasons — internal
packages export only to cross the internal/ boundary and are not public API, and
store backends implement a Store interface documented once on the interface,
where a per-backend copy would be a fourth place to drift.

REAL DOC BUGS (11), which is what made this worth doing rather than suppressing:

  - SIX orphaned doc comments had drifted onto the wrong declaration, so the
    function they describe had NO documentation and the function above them was
    documented twice. Incoming, NodesAsOf, IncomingRelationships (badger and
    memory), ForEachNodeByLabelPropertyRange and ErrForeignStampImplausible are
    each moved back to their own declaration, verified undocumented first.
  - FIVE stale names left by renames: DropProperty/DropTemporal/
    DropHighFrequency/DropVector documented methods now called Delete*, and
    GetByIDs documented GetNodesByIDs.

DEAD CODE (10): nine functions and one const with ZERO call sites anywhere —
verified by grepping for call syntax, not just the name. lockActiveCoreContext,
resolve's ready, appendOpsLogged, logCascadeNodeDelete, truncateHistoryByPrefix,
trimHistoryFromPrefix, and memory's three history-log helpers. Uncalled code is
the anti-pattern lessons.md 26 names; deleting it is the fix.

SIMPLIFICATIONS the linter was right about (6). emit in both history scanners
appends to a slice and cannot fail, so its always-nil error is gone — and that
CASCADED, because drainPendingBefore could then no longer fail either. Same for
collectLogShards. seedForeverOwnersFromCurrent took a label it never read.

DELIBERATE PATTERNS kept, with narrow //nolint and a reason rather than a
rewrite that would destroy the assertion: the pointer-identity checks
(g.Nodes() != g.Nodes() IS the test), passing a nil context to assert
ErrNilContext, and an empty critical section asserting a writer acquires
immediately.

Verified: build, vet and gofmt clean; golangci-lint 0 issues; 38/38 packages
pass; coverage gate 85.5% >= 80%.

+++ pkg/graph/store/sharded/bench_lanes_test.go
@@ -161,7 +161,7 @@ func BenchmarkIngestShardedLanesBulk(b *testing.B) {
 func BenchmarkIngestShardedLanesBulkScaling(b *testing.B) {
 	for _, lanes := range []uint8{2, 4, 8, 16} {
 		b.Run(fmt.Sprintf("lanes-%d", lanes), func(b *testing.B) {
-			st, err := sharded.New(sharded.Config{InMemory: true, BaseSlot: 0, SlotCount: 2 + uint8(lanes)})
+			st, err := sharded.New(sharded.Config{InMemory: true, BaseSlot: 0, SlotCount: 2 + lanes})
 			if err != nil {
 				b.Fatal(err)
 			}
